boards/esp32: Refactor SPI hooks to be shared among ESP32 devkits
