+++ boards/xtensa/esp32/common/src/Make.defs
@@ -30,6 +30,10 @@ ifeq ($(CONFIG_I2C_DRIVER),y)
   CSRCS += esp32_board_i2c.c
 endif
 
+ifeq ($(CONFIG_ESP32_SPI),y)
+  CSRCS += esp32_board_spi.c
+endif
+
 ifeq ($(CONFIG_ESP32_WIRELESS),y)
   CSRCS += esp32_board_wlan.c
 endif
