--- boards/xtensa/esp32/esp32-devkitc/src/esp32_spi.c
@@ -1,67 +1,0 @@
-/****************************************************************************
- * boards/xtensa/esp32/esp32-devkitc/src/esp32_spi.c
- *
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.  The
- * ASF licenses this file to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance with the
- * License.  You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
- * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.  See the
- * License for the specific language governing permissions and limitations
- * under the License.
- *
- ****************************************************************************/
-
-/****************************************************************************
- * Included Files
- ****************************************************************************/
-
-#include <nuttx/config.h>
-
-#include <stdint.h>
-#include <stdbool.h>
-#include <debug.h>
-
-#include <nuttx/spi/spi.h>
-#include <arch/board/board.h>
-
-#include "esp32-devkitc.h"
-
-/****************************************************************************
- * Public Functions
- ****************************************************************************/
-
-/****************************************************************************
- * Name:  esp32_spi2_status
- ****************************************************************************/
-
-uint8_t esp32_spi2_status(FAR struct spi_dev_s *dev, uint32_t devid)
-{
-  uint8_t status = 0;
-
-#ifdef CONFIG_MMCSD_SPI
-  if (devid == SPIDEV_MMCSD(0))
-    {
-       status |= SPI_STATUS_PRESENT;
-    }
-#endif
-
-  return status;
-}
-
-/****************************************************************************
- * Name:  esp32_spi3_status
- ****************************************************************************/
-
-uint8_t esp32_spi3_status(FAR struct spi_dev_s *dev, uint32_t devid)
-{
-  uint8_t status = 0;
-
-  return status;
-}
