+++ boards/xtensa/esp32/esp32-wrover-kit/src/Make.defs
@@ -39,10 +39,6 @@ CSRCS += esp32_reset.c
 endif
 endif
 
-ifeq ($(CONFIG_ESP32_SPI),y)
-CSRCS += esp32_spi.c
-endif
-
 ifeq ($(CONFIG_MMCSD),y)
 CSRCS += esp32_mmcsd.c
 endif
